[REL-8210] account for subdirectory in ref filepaths (#516)
We were inadverntently ignoring the `subdirectory` option when creating
`File` reps, which led to the tool searching zero files when using the
project-level `dir` property in `.launchdarkly/coderefs.yml` (There is a
workaround to just not specify the `dir` property, but this leads to
broken links in gonfalon). The reason for this is that when using the
`subdirectory` property, the `workspace` (this is the root directory
that `ld-find-code-refs` searches under) becomes `<root>/subdirectory`,
but the project-level `dir` property is relative to the repo root, so
the tool would end up looking in `<root>/subdirectory/subdirectory` for
files to scan which obviously doesn't exist. The fix here is to remove
the `subdirectory` (if it exists) from the `file`'s `path` property.
Existing tests pass; added new tests for the `subdirectory` use-case.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-8210: Code Refs - Using subdirectory flag
creates broken link in LD UI when clicking View
source](https://launchdarkly.atlassian.net/browse/REL-8210)
<!-- end-ld-jira-link -->