[REL-4161] fix syntax error; add a bit more logging (#522)
Hopefully this will be the last one 😅.
Yesterday evening I added the CHANGELOG update which I had overlooked
initially because Releaser did it behind-the-scenes. But there was an
extra `)` in the code 🤦, so that's fixed here, and I'm also logging out
the changelog so future releasers will be able to see the diff and the
updated CHANGELOG when they run a dry run or actual release. There's
also a minor refactor here. It's nothing big, but it makes the updates a
bit harder to grok, so I'll try to comment in the code to explain what
was changed.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->