chore: [REL-4161] thread homebrew-tap deploy key through to goreleaser (#547)
I think we need to add the `LAUNCHDARKLY_HOMEBREW_TAP_DEPLOY_KEY` as an
env var in order to make it available to goreleaser. I think the reason
this didn't work with the `ssh_key_by_repo` action is because the GHA
itself is not accessing the homebrew-tap repo; it's goreleaser that's
trying to make changes there. However, since we added `homebrew-tap` as
a `cross_repo_deploy_key` for `ld-find-code-refs`, I believe the secret
lives on the repo (see the description in screenshot), so I'm hopeful
grabbing it from `secrets` and setting as an env var will be enough to
give access to goreleaser.

<img width="1579" height="240" alt="Screenshot 2025-08-04 at 1 29 55 PM"
src="https://github.com/user-attachments/assets/5131563e-f3c4-42ab-b58a-50322665fb64"
/>

<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->