chore: [REL-4161] update goreleaser config for version 1.20 (#549)
Apparently, the config properties have changed over time, and
`directory` used to be called `folder`. Docs for GoReleaser 1.20 can be
found
[here](https://github.com/goreleaser/goreleaser/blob/v1.20.0/www/docs/customization/homebrew.md).
Eventually, it would be good to update the Goreleaser version here, but
for now I just want to cut a new release as it's blocking a fair number
of improvements, some of which are kinda important.

I had also removed the `ids` property, but I'm adding it back in here. I
don't really understand what it's for. The docs say: `IDs of the
archives to use`. not sure what that means, but seems like leaving this
in for now would make the most sense, since it has existed up until this
point.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->