chore: [REL-4161] use printf to ensure we get the exact value for homebrew key (#554)
The saga continues. Last time I got an
[error](https://github.com/launchdarkly/ld-find-code-refs/actions/runs/16777430844/job/47506788949#step:7:633)
that is likely related to a malformed homebrew-tap deploy key. Giving
`printf`, and no json encoding a shot. Hopefully, this will write the
unaltered value of the key to the file and we'll be good to go.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->