chore: [REL-4161] dont run setup or cleanup if tag already exists (#557)
I had setup some idempotency checks in the GHA, Bitbucket Pipe and
Circleci Orb scripts, so that if we had a partially successful run, we
would skip any releases that had already completed successfully (based
on the existence of the tag). However, I realized that the setup for
those releases also needs to be skipped if the tag exists, so that's
what this update does.

Cursor also noticed that using the fully-qualified repo in the tag check
is wiser since we have this weird nested git repo thing going on, and we
don't want to accidentally check the wrong repo. I'll be updating this
so the repos are fully isolated once I get this release out the door.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->