chore: [REL-4161] fix up tag checks (#558)
Two things being updated here:
1. update the tag-check (for this repo) to only check the remote repo
for the tag. It will definitely exist in the local copy since we need to
create it before running Goreleaser. Other minor tweak is that instead
of bailing on the entire action, we just return 0, so the action can
continue to run and complete the final step of creating the Github
release. The only reason we're likely to be running the action again if
the tag was created, is that the Github release failed, so we want to
let the action continue, so it can do that last step.
2. checking for existing tag (GHA and Bitbucket Pipe) on dry-run as
well. This was fixed for real runs of the action, but dry-runs will also
bomb out if the tag exists (the tag could exist if we previously had a
partially successful run, but need to re-run the action to retry the
parts that didn't complete successfully)
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->