chore: [REL-4161] make the circleci orb check more reliable (#559)
Orb check caused the action to
[fail](https://github.com/launchdarkly/ld-find-code-refs/actions/runs/16807638943/job/47604213552#step:7:679)
last time. The way we were doing the check was not great as it pulls
every single public Orb and greps for the coderefs orb with the version
we're looking for. I tried that locally and verified that it doesn't
find the orb (even though it exists). I also tried the updated orb check
locally, which is way more efficient and does find the Orb we're looking
for.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->