chore: [REL-4161] use elevated token for git stuff (#562)
The GITHUB_TOKEN that's built-in to the GHA is not authorized to push
directly to main. We need to do this in order to push up our committed
changes to the remote repo so we can create the release. I had a new SSM
parameter added that will allow this, so we should be able to cut the
release now with this new token. I needed to get a little creative since
the value is not known until after the `get secrets` step runs, but I
think what I have here should work.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-4161: Migrate ld-find-code-refs from Releaser
to GHA](https://launchdarkly.atlassian.net/browse/REL-4161)
<!-- end-ld-jira-link -->