feat: [REL-8033] Prepare release of version 2.14.0 (#563)
Our release.yml GHA has managed to successfully release the Homebrew
Formula and our three CI coderefs tools (Github Action, CircleCI Orb and
BitBucket Pipe), but it's consistently having issues with the Github
release. Those newly-released CI tools point to the `v2.14.0` tag of
this repo, which doesn't exist yet. So there is some urgency to get the
Github release out as quickly as we can. Since we can't seem to get
around the branch protection issues we've run into, we've decided to
release version 2.14.0 manually, and then rework the GHA to make use of
Release Please, which doesn't require pushes to the main branch. This PR
will prepare the necessary updates for the release, and I'll create the
binaries locally and create the tag and the release manually once this
is merged.
<!-- ld-jira-link -->
---
Related Jira issue: [REL-8033: Release a new version of
`ld-find-code-refs` and `find-code-references`
GHA](https://launchdarkly.atlassian.net/browse/REL-8033)
<!-- end-ld-jira-link -->